Adopted single SplatC over existing *Const8 instructions
This came from a tangent in exploring adding const immediates to
all binary instructions in the form of a splatted-sign-extended u8
value where the register would normally go.

Mostly due to time, extra complexity, and increased number of
instructions, I've decided to not explore this for now, though in
theory it would have an improvement in both performance (number of
instructions) and register usage.

But one interesting observation the fell out was that the
extend/splating const8 instructions were mostly redundant. The size
of the const field is by fixed to 8-bits, so logically the width + lane
npw2 fields could encode a splatted-signed-extended value with no
extra encoding space.

Just to clarify what I mean by splatted-signed-extended, this:

u32x4.splat_c r1, 0xfc

Would be the same as this:

u128.move_const r1, 0xfffffffcfffffffcfffffffcfffffffc

This is quite a powerful encoding, allowing any number from -128 to 127
to be encoded and splatted across all lanes for any register geometry
in a single instruction

---

It's worth noting this is a small reduction in capability, as the
extend_const8_u doesn't have a direct replacement if the value is >=0x80
and type >=u16. But this is minor, and can be replace by the more
complete extend_const_u instruction.

Also the compiler is currently limited to outputing const splats OR
const extend_s, so the full range can't be leveraged very easily at the
moment.