Tried to reduce clones a bit more
- Changed declassify-functions to take a reference
- Changed compiled lambdas to take references
- Change closure to keep the working state allocated (we're not thread
  safe remember!)

Unfortunately the current small-object optimization does not work well
with the take-everything-by-value type signature. The reason is that in
order for values to be shared between operations, their Rcs need to be
cloned outside of the operation definitions themselves. This means
operations that could avoid copies (such as extract) are forced to go
through the tree-alloc->compile->exec loop.

Currently this doesn't seem like an easy fix, but it can at least be
alleviated by taking declassifies by reference, which removes an extra
alloc->ocmpile->exec loop from declassifies, calling compiled functions,
etc.