Adopted a 64-bit instruction set
[4|4|- 8 -|--  16  --|--  16  --|--  16  --]
 ^ ^   ^        ^          ^          ^-- 16-bit ra
 | |   |        |          '------------- 16-bit rb
 | |   |        '------------------------ 16-bit rd
 | |   '--------------------------------- 8-bit opcode
 | '------------------------------------- 4-bit npw2(lanes)
 '--------------------------------------- 4-bit npw2(size)

Suggested by theotherjimmy, this aggressive encoding solves a number
of problems with the previous 32-bit instruction set. The main
motivation being that the extra bit in both the npw2 and lnpw2 field
will allow for integer sizes all the way up to u262144 instead of the
previous u512 limitation.

The obvious downside is a ~2x larger bytecode size. However, since all
loops are forcibly unrolled, fitting into the smaller caches may not be
as important as the decoding complexity. So this impact may not be that
significant (or maybe it will be, we'll just have to see).

64-bit instruction set benefits:
- Increased integer width from u512 -> u262144
- Increased registers from 256 -> 65536, should reduce register overflows
- Simpler decode, 8-bit opcode alone determines operation
- Almost all instructions are 3-reg (rd=ra+rb), reducing extra moves
- Triops (shuffle, select, replace) no longer need special encoding, and
  we can add more
- splat_c increases its immediate size from a u8 to a u32, this allows
  much larger immediate, and since the next immediate size would
  naturally be a u64, this avoids some annoyances with alignment in the
  instruction stream