Implemented in-place multiplication
By multiplying top-to-bottom and inward, it turns out we can perform
arbitrary precision multiplication in-place. In-place in the context
meaning any and both arguments can overlap with the destination register
without getting corrupted mid-multiplication.

I haven't found an algorithm that does this yet without requiring some
external allocation, there seem to be a common top-to-bottom multiply
algorithm but it requires O(log n) extra space. With memory allocations
the overhead of the allocation itself is the main cost, so this doesn't
actually buy much in this case.

Maybe it's because this multiplication is asymptotically less efficient,
requiring O(n^3) additions and O(n^2) multiplications unlike the O(n^2)
additions/multiplications of other algorithms.

Is this faster than allocating extra space? Note sure. It would be
interesting to compare performance at some point.

---

Note, it is possible switch between a more efficient multiplication
algorithm when registers do not overlap (register ordering does not
depend on secret info), but as a toy project, two algorithms are not
worth maintaining right now.