Reorganized shuffle/select arguments to allow in-place operations
On paper, shuffle looks like it can never allow in-place operations,
since each lane in the destination can depend on any lane from either
source, we can't allow any overwriting to occur until all lanes have
been visited.

But! shuffle and select are unique in that they read from 3 registers
for their operation, and one register, the index register, is trivially
updatable in-place.

The solution is to just reorganize the shuffle arguments such that the
index register is the same as the destination register. Interesting to
note that this was the original argument order, before the the registers
were moved to reduce extra register copies. Though in this case, the
extra register copies are what enables in the in-place shuffle.