Allowed for more direct data movement in extract/replace/reverse
This is an optimization to avoid dropping down into the secret-u-engine
for simple data movements operations, since these tend to happen very
heavily as a part of data packing/unpacking in order to move data into
secret-u computations. This shouldn't violated any constant-time
guarantees since the only operations are data movement, which is already
unavoidable when constructing the initial values. It does mean possibly
move movement of the data, but this may be worth it for the peformance
improvement. And now all of the extra work to avoid packing/unpack
before passing values to lazy_compiled functions can be avoided.

Note this doesn't change splat. Splat could be modified to construct
a new immediate value, but it's usually intended for compressed
constants, so constructing a new immediate value will usually be a waste
of effort and may even reduce performance due to increased immediate
sizes.